Module aws_sdk_computeoptimizer::model

Modules

account_enrollment_status

See AccountEnrollmentStatus

auto_scaling_group_configuration

See AutoScalingGroupConfiguration

auto_scaling_group_recommendation

See AutoScalingGroupRecommendation

auto_scaling_group_recommendation_option

See AutoScalingGroupRecommendationOption

ebs_filter

See EbsFilter

ebs_utilization_metric

See EbsUtilizationMetric

enrollment_filter

See EnrollmentFilter

export_destination

See ExportDestination

filter

See Filter

get_recommendation_error

See GetRecommendationError

instance_recommendation

See InstanceRecommendation

instance_recommendation_option

See InstanceRecommendationOption

job_filter

See JobFilter

lambda_function_memory_projected_metric

See LambdaFunctionMemoryProjectedMetric

lambda_function_memory_recommendation_option

See LambdaFunctionMemoryRecommendationOption

lambda_function_recommendation

See LambdaFunctionRecommendation

lambda_function_recommendation_filter

See LambdaFunctionRecommendationFilter

lambda_function_utilization_metric

See LambdaFunctionUtilizationMetric

projected_metric

See ProjectedMetric

reason_code_summary

See ReasonCodeSummary

recommendation_export_job

See RecommendationExportJob

recommendation_preferences

See RecommendationPreferences

recommendation_source

See RecommendationSource

recommendation_summary

See RecommendationSummary

recommended_option_projected_metric

See RecommendedOptionProjectedMetric

s3_destination

See S3Destination

s3_destination_config

See S3DestinationConfig

summary

See Summary

utilization_metric

See UtilizationMetric

volume_configuration

See VolumeConfiguration

volume_recommendation

See VolumeRecommendation

volume_recommendation_option

See VolumeRecommendationOption

Structs

AccountEnrollmentStatus

Describes the enrollment status of an organization's member accounts in Compute Optimizer.

AutoScalingGroupConfiguration

Describes the configuration of an Auto Scaling group.

AutoScalingGroupRecommendation

Describes an Auto Scaling group recommendation.

AutoScalingGroupRecommendationOption

Describes a recommendation option for an Auto Scaling group.

EbsFilter

Describes a filter that returns a more specific list of Amazon Elastic Block Store

EbsUtilizationMetric

Describes a utilization metric of an Amazon Elastic Block Store (Amazon EBS)

EnrollmentFilter

Describes a filter that returns a more specific list of account enrollment statuses.

ExportDestination

Describes the destination of the recommendations export and metadata files.

Filter

Describes a filter that returns a more specific list of recommendations. Use this

GetRecommendationError

Describes an error experienced when getting recommendations.

InstanceRecommendation

Describes an Amazon EC2 instance recommendation.

InstanceRecommendationOption

Describes a recommendation option for an Amazon EC2 instance.

JobFilter

Describes a filter that returns a more specific list of recommendation export jobs.

LambdaFunctionMemoryProjectedMetric

Describes a projected utilization metric of an Lambda function

LambdaFunctionMemoryRecommendationOption

Describes a recommendation option for an Lambda function.

LambdaFunctionRecommendation

Describes an Lambda function recommendation.

LambdaFunctionRecommendationFilter

Describes a filter that returns a more specific list of Lambda

LambdaFunctionUtilizationMetric

Describes a utilization metric of an Lambda function.

ProjectedMetric

Describes a projected utilization metric of a recommendation option, such as an

ReasonCodeSummary

A summary of a finding reason code.

RecommendationExportJob

Describes a recommendation export job.

RecommendationPreferences

Describes preferences for recommendations.

RecommendationSource

Describes the source of a recommendation, such as an Amazon EC2 instance or

RecommendationSummary

A summary of a recommendation.

RecommendedOptionProjectedMetric

Describes a projected utilization metric of a recommendation option.

S3Destination

Describes the destination Amazon Simple Storage Service (Amazon S3) bucket name and

S3DestinationConfig

Describes the destination Amazon Simple Storage Service (Amazon S3) bucket name and

Summary

The summary of a recommendation.

UtilizationMetric

Describes a utilization metric of a resource, such as an Amazon EC2

VolumeConfiguration

Describes the configuration of an Amazon Elastic Block Store (Amazon EBS)

VolumeRecommendation

Describes an Amazon Elastic Block Store (Amazon EBS) volume recommendation.

VolumeRecommendationOption

Describes a recommendation option for an Amazon Elastic Block Store (Amazon EBS)

Enums

CpuVendorArchitecture

EbsFilterName

EbsFinding

EbsMetricName

EnrollmentFilterName

ExportableAutoScalingGroupField

ExportableInstanceField

ExportableLambdaFunctionField

ExportableVolumeField

FileFormat

FilterName

Finding

FindingReasonCode

InstanceRecommendationFindingReasonCode

JobFilterName

JobStatus

LambdaFunctionMemoryMetricName

LambdaFunctionMemoryMetricStatistic

LambdaFunctionMetricName

LambdaFunctionMetricStatistic

LambdaFunctionRecommendationFilterName

LambdaFunctionRecommendationFinding

LambdaFunctionRecommendationFindingReasonCode

MetricName

MetricStatistic

PlatformDifference

RecommendationSourceType

ResourceType

Status