Struct aws_sdk_computeoptimizer::model::lambda_function_recommendation::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for LambdaFunctionRecommendation

Implementations

impl Builder

pub fn function_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the current function.

pub fn set_function_arn(self, input: Option<String>) -> Self

pub fn function_version(self, input: impl Into<String>) -> Self

The version number of the current function.

pub fn set_function_version(self, input: Option<String>) -> Self

pub fn account_id(self, input: impl Into<String>) -> Self

The Amazon Web Services account ID of the function.

pub fn set_account_id(self, input: Option<String>) -> Self

pub fn current_memory_size(self, input: i32) -> Self

The amount of memory, in MB, that's allocated to the current function.

pub fn set_current_memory_size(self, input: Option<i32>) -> Self

pub fn number_of_invocations(self, input: i64) -> Self

The number of times your function code was applied during the look-back period.

pub fn set_number_of_invocations(self, input: Option<i64>) -> Self

pub fn utilization_metrics(
self,
input: impl Into<LambdaFunctionUtilizationMetric>
) -> Self

pub fn set_utilization_metrics(
self,
input: Option<Vec<LambdaFunctionUtilizationMetric>>
) -> Self

pub fn lookback_period_in_days(self, input: f64) -> Self

The number of days for which utilization metrics were analyzed for the function.

pub fn set_lookback_period_in_days(self, input: Option<f64>) -> Self

pub fn last_refresh_timestamp(self, input: Instant) -> Self

The timestamp of when the function recommendation was last refreshed.

pub fn set_last_refresh_timestamp(self, input: Option<Instant>) -> Self

pub fn finding(self, input: LambdaFunctionRecommendationFinding) -> Self

The finding classification of the function.

Findings for functions include:

Optimized — The function is correctly provisioned to run your workload based on its current configuration and its utilization history. This finding classification does not include finding reason codes.

NotOptimized — The function is performing at a higher level (over-provisioned) or at a lower level (under-provisioned) than required for your workload because its current configuration is not optimal. Over-provisioned resources might lead to unnecessary infrastructure cost, and under-provisioned resources might lead to poor application performance. This finding classification can include the MemoryUnderprovisioned and MemoryUnderprovisioned finding reason codes.

Unavailable — Compute Optimizer was unable to generate a recommendation for the function. This could be because the function has not accumulated sufficient metric data, or the function does not qualify for a recommendation. This finding classification can include the InsufficientData and Inconclusive finding reason codes.

Functions with a finding of unavailable are not returned unless you specify the filter parameter with a value of Unavailable in your GetLambdaFunctionRecommendations request.

pub fn set_finding(
self,
input: Option<LambdaFunctionRecommendationFinding>
) -> Self

pub fn finding_reason_codes(
self,
input: impl Into<LambdaFunctionRecommendationFindingReasonCode>
) -> Self

pub fn set_finding_reason_codes(
self,
input: Option<Vec<LambdaFunctionRecommendationFindingReasonCode>>
) -> Self

pub fn memory_size_recommendation_options(
self,
input: impl Into<LambdaFunctionMemoryRecommendationOption>
) -> Self

pub fn set_memory_size_recommendation_options(
self,
input: Option<Vec<LambdaFunctionMemoryRecommendationOption>>
) -> Self

pub fn build(self) -> LambdaFunctionRecommendation

Consumes the builder and constructs a LambdaFunctionRecommendation