Module aws_sdk_computeoptimizer::operation

Structs

DescribeRecommendationExportJobs

Describes recommendation export jobs created in the last seven days.

ExportAutoScalingGroupRecommendations

Exports optimization recommendations for Auto Scaling groups.

ExportEBSVolumeRecommendations

Exports optimization recommendations for Amazon EBS volumes.

ExportEC2InstanceRecommendations

Exports optimization recommendations for Amazon EC2 instances.

ExportLambdaFunctionRecommendations

Exports optimization recommendations for Lambda functions.

GetAutoScalingGroupRecommendations

Returns Auto Scaling group recommendations.

GetEBSVolumeRecommendations

Returns Amazon Elastic Block Store (Amazon EBS) volume recommendations.

GetEC2InstanceRecommendations

Returns Amazon EC2 instance recommendations.

GetEC2RecommendationProjectedMetrics

Returns the projected utilization metrics of Amazon EC2 instance

GetEnrollmentStatus

Returns the enrollment (opt in) status of an account to the Compute Optimizer

GetEnrollmentStatusesForOrganization

Returns the Compute Optimizer enrollment (opt-in) status of organization member

GetLambdaFunctionRecommendations

Returns Lambda function recommendations.

GetRecommendationSummaries

Returns the optimization findings for an account.

UpdateEnrollmentStatus

Updates the enrollment (opt in and opt out) status of an account to the Compute Optimizer service.